Update LabExercise2.ipynb
SMOTE (Synthetic Minority Over-sampling Technique) is used to handle class imbalance in the training data.

Random Forest is used for feature selection, and features with importance greater than a certain threshold are selected.

Logistic Regression is trained on the selected features with hyperparameters tuned using Randomized Search.

Performance metrics are evaluated on the test set, and overfitting is checked by comparing training and test set scores.